--- invite/src/types.ts
@@ -1,44 +1,0 @@
-/**
- * Copyright 2020 The AMP HTML Authors. All Rights Reserved.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS-IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-/**
- * A standard logging interface.
- */
-export interface Logger {
-  debug(message: string, ...extraInfo: any[]): void;
-  warn(message: string, ...extraInfo: any[]): void;
-  error(message: string, ...extraInfo: any[]): void;
-  info(message: string, ...extraInfo: any[]): void;
-}
-
-/**
- * Possible invite action types.
- */
-export enum InviteAction {
-  INVITE = 'invite',
-  INVITE_AND_ASSIGN = 'invite_and_assign',
-}
-
-/**
- * An invite triggered by the bot.
- */
-export interface Invite {
-  username: string;
-  repo: string;
-  issue_number: number;
-  action: InviteAction;
-  archived?: boolean;
-}
