+++ invite/types/invite-bot.d.ts
@@ -1,0 +1,43 @@
+/**
+ * Copyright 2020 The AMP HTML Authors. All Rights Reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS-IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+declare module 'invite-bot' {
+  /**
+   * A standard logging interface.
+   */
+  export interface Logger {
+    debug(message: string, ...extraInfo: unknown[]): void;
+    warn(message: string, ...extraInfo: unknown[]): void;
+    error(message: string, ...extraInfo: unknown[]): void;
+    info(message: string, ...extraInfo: unknown[]): void;
+  }
+
+  /**
+   * Possible invite action types.
+   */
+  export type InviteActionType = 'invite' | 'invite_and_assign';
+
+  /**
+   * An invite triggered by the bot.
+   */
+  export interface Invite {
+    username: string;
+    repo: string;
+    issue_number: number;
+    action: InviteAction;
+    archived?: boolean;
+  }
+}
